Add docker (#18)
This commit adds support for using docker in racketRun. This feature can be enabled by setting the environment variable GRAPHREDEX_DOCKER to 1. 

To A docker image based on ubuntu is used because we could not get racket working in alpine. 

The xvfb within the docker image is old and requires a patch to separate stdout and stderr.

List of squashed commits:

* docker start

* WIP: needs bodges...

* Add ENV var to say where the DB is

This is needed to execute the server within a docker container.

* fixes to get it working

* send right environment

* patch xvfb

* optimise

* note on security