fix: ForceShower went nuts when toggled from TreeShower
The was caused by the initial value of the heat in the system.

In the normal use case the ForceShower is constructed and the heat is
set to one. As time progresses this decreases (without effect because
there are no nodes). When an example was selected, the nodes were added
to a cold system. When switching from TreeShower, a new ForceShower is
constructed and immediately used (hot). This caused some wacko
behaviour.